Integração com o jwt e Swagger funcionando com authorization

+++ naum-spring-api/src/main/java/school/sptech/naumspringapi/dto/barbeariaDto/BarbeariaListagemDto.java
@@ -2,10 +2,11 @@ package school.sptech.naumspringapi.dto.barbeariaDto;
 
 import lombok.Data;
 import school.sptech.naumspringapi.dto.enderecoDto.EnderecoListagemDto;
+import school.sptech.naumspringapi.entity.Endereco;
 
 @Data
 public class BarbeariaListagemDto {
-    private int id;
+    private Long id;
     private String nome;
     private String linkBarbearia;
     private String fotoBarbearia;
