+++ naum-spring-api/src/main/java/school/sptech/naumspringapi/entity/Endereco.java
@@ -1,8 +1,8 @@
 package school.sptech.naumspringapi.entity;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import jakarta.persistence.*;
 
 @Entity
 @Getter
@@ -10,12 +10,11 @@ import jakarta.persistence.*;
 public class Endereco {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String cidade;
     private String cep;
     private String numero;
     private String bairro;
     private String uf;
     private String rua;
-
 }
