+++ naum-spring-api/src/main/java/school/sptech/naumspringapi/repository/EnderecoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import school.sptech.naumspringapi.entity.Endereco;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
 }
