+++ naum-spring-api/src/main/java/school/sptech/naumspringapi/repository/LoginAdmRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import school.sptech.naumspringapi.entity.LoginAdm;
 
 @Repository
-public interface LoginAdmRepository extends JpaRepository<LoginAdm, Integer> {
+public interface LoginAdmRepository extends JpaRepository<LoginAdm, Long> {
 }
